+++ test/controllers/patients_controller_test.rb
@@ -6,51 +6,6 @@ class PatientsControllerTest < ActionController::TestCase
     login(users(:screener))
   end
 
-  test "should get patient stickies" do
-    app = proc do |env|
-      [200, { 'Content-Type' => 'application/json' }, [ [{ id: 1, description: "6-month followup call", all_day: true, completed: false, tags: [ { id: 1, name: "Phone Call", color: '#aa00bb' } ] },
-                                                         { id: 2, description: "8-month visit", all_day: true, due_date: "2011-04-11T14:45:00-04:00", completed: false, tags: [ { id: 2, name: "Visit", color: '#330077' } ] } ].to_json ]]
-    end
-    Artifice.activate_with(app) do
-      post :stickies, id: patients(:with_subject_code), format: 'js'
-    end
-    assert_not_nil assigns(:stickies)
-    # Can't assert this if the task tracker server is down or not set
-    unless TASK_TRACKER_URL.blank? or TT_EMAIL.blank? or TT_PASSWORD.blank?
-      assert_equal 1, assigns(:stickies).first['id']
-      assert_equal "6-month followup call", assigns(:stickies).first['description']
-      assert_equal ({ display: 'No Date', id: 'no_date' }), assigns(:stickies).first['month_year']
-      assert_equal ({ display: 'April 2011', id: '042011' }), assigns(:stickies).last['month_year']
-    end
-    assert_template 'stickies'
-    assert_response :success
-  end
-
-  test "should fail gracefully on unknown server response for task tracker stickies" do
-    app = proc do |env|
-      [500, { 'Content-Type' => 'text/html' }, [ "Internal Server Error" ]]
-    end
-    Artifice.activate_with(app) do
-      post :stickies, id: patients(:with_subject_code), format: 'js'
-    end
-    assert_not_nil assigns(:stickies)
-    assert_equal [], assigns(:stickies)
-    assert_template 'stickies'
-    assert_response :success
-  end
-
-  test "should not show patient stickies without subject code to subject handler" do
-    login(users(:subject_handler))
-    app = proc do |env|
-      [200, { 'Content-Type' => 'application/json' }, [ [{ id: 1, description: "6-month followup call", all_day: true, completed: false, tags: [ { id: 1, name: "Phone Call", color: '#aa00bb' } ] },
-                                                         { id: 2, description: "8-month visit", all_day: true, due_date: "2011-04-11T14:45:00-04:00", completed: false, tags: [ { id: 2, name: "Visit", color: '#330077' } ] } ].to_json ]]
-    end
-    Artifice.activate_with(app) do
-      post :stickies, id: patients(:without_subject_code), format: 'js'
-    end
-    assert_response :success
-  end
-
   test "should get index" do
     get :index
     assert_response :success